Rename WifiConnector component
Component named after what it actually does, configuring the wifi. The
namespace is now in the following convention:

  src/(<namespace>/)*/file

using the includedir only on src, the namespace is visible in the
includes in the other files.
