+++ components/wifi_configurator/component.mk
